// dpbloom-client/src/app/core/api/model/createCourseDto.ts
/**
 * DPBloom | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateCourseDto { 
    title?: string;
    description?: string | null;
    isPublished?: boolean;
}